refactor(studio): extract LintModal, MediaPreview, shared media regexes from App.tsx
- Extract LintModal (130 lines) to components/LintModal.tsx
- Extract MediaPreview (75 lines) to components/MediaPreview.tsx
- Extract shared media regexes to utils/mediaTypes.ts (single source
  of truth — App.tsx and AssetsTab.tsx were silently diverged)
- Add useMemo for compositions/assets derivation and FileTree buildTree
- Debounce handleContentChange PUT (600ms, was firing every keystroke)
- Guard projectId null in LintModal render
